requirements.txt
langchain-community==0.0.38

crewai>=1.8.0
crewai-tools>=0.2.0
langchain-openai>=1.1.0
qdrant-client>=1.16.0
firecrawl-py>=4.13.0
PyPDF2>=3.0.1
python-dotenv>=1.1.0

# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# LlamaIndex Dependencies
llama-index>=0.9.48
llama-index-vector-stores-qdrant>=0.2.8
llama-index-embeddings-openai>=0.1.6
llama-index-llms-openai>=0.1.6

# Additional Document Processing
python-docx>=1.1.0

# crewai==0.28.8
# crewai-tools==0.2.6
# langchain-community==0.0.38
# qdrant-client==1.7.0
# firecrawl-py==0.0.16
# PyPDF2==3.0.1
